Move ade functions into own files
