+++ code/globalincs/windebug.cpp
@@ -32,7 +32,7 @@
 #include "globalincs/pstypes.h"
 #include "globalincs/systemvars.h"
 #include "cmdline/cmdline.h"
-#include "parse/lua.h"
+#include "scripting/lua.h"
 #include "parse/parselo.h"
 #include "debugconsole/console.h"
