+++ code/scripting/lua.cpp
@@ -40,9 +40,9 @@
 #include "network/multimsgs.h"
 #include "object/objectshield.h"
 #include "object/waypoint.h"
-#include "lua.h"
+#include "scripting/lua.h"
 #include "parse/parselo.h"
-#include "scripting.h"
+#include "scripting/scripting.h"
 #include "particle/particle.h"
 #include "playerman/player.h"
 #include "render/3d.h"
@@ -58,163 +58,7 @@
 #define BMPMAN_INTERNAL
 #include "bmpman/bm_internal.h"
 
-//*************************Lua globals*************************
-SCP_vector<ade_table_entry> Ade_table_entries;
-
-//*************************Lua classes************************
-
-//Library class
-//This is what you define a variable of to make new libraries
-class ade_lib : public ade_lib_handle {
-public:
-	ade_lib(char *in_name, ade_lib_handle *parent=NULL, const char *in_shortname=NULL, const char *in_desc=NULL) {
-		ade_table_entry ate;
-
-		ate.Name = in_name;
-		ate.ShortName = in_shortname;
-		ate.Instanced = true;
-
-		//WMC - Here's a little hack.
-		//Lua did not work with __len on standard table objects.
-		//So instead, all FS2 libraries are now userdata.
-		//This means that no new functions can be added from
-		//within the scripting environment, but I don't think
-		//there will be any catastrophic consequences.
-		ate.Type = 'o';
-		ate.Value.Object.idx = Ade_table_entries.size();
-		ate.Value.Object.sig = NULL;
-		ate.Value.Object.buf = &Num_reinforcements;	//WMC - I just chose Num_ship_classes randomly. MageKing17 - changed to Num_reinforcements, likewise at random, due to the removal of Num_ship_classes
-		ate.Value.Object.size = sizeof(Num_reinforcements);
-		ate.Description = in_desc;
-
-		if(parent != NULL)
-			Ade_table_entries[parent->GetIdx()].AddSubentry(ate);
-		else
-			Ade_table_entries.push_back(ate);
-		LibIdx = Ade_table_entries.size()-1;
-	}
-
-	const char *GetName();
-};
-
-const char *ade_lib::GetName()
-{
-	if(GetIdx() == UINT_MAX)
-		return "<Invalid>";
-
-	return Ade_table_entries[GetIdx()].GetName();
-}
-
-//Function helper class
-//Lets us add functions via its constructor
-class ade_func : public ade_lib_handle {
-public:
-	ade_func(const char *name, lua_CFunction func, ade_lib_handle &parent, const char *args=NULL, const char *desc=NULL,
-			 const char *ret_type=NULL, const char*ret_desc=NULL) {
-		ade_table_entry ate;
-
-		ate.Name = name;
-		ate.Instanced = true;
-		ate.Type = 'u';
-		ate.Value.Function = func;
-		ate.Arguments = args;
-		ate.Description = desc;
-		ate.ReturnType = ret_type;
-		ate.ReturnDescription = ret_desc;
-
-		Ade_table_entries[parent.GetIdx()].AddSubentry(ate);
-		LibIdx = Ade_table_entries.size()-1;
-	}
-};
-
-class ade_virtvar : public ade_lib_handle {
-public:
-	ade_virtvar(const char *name, lua_CFunction func, ade_lib_handle &parent, const char *args=NULL, const char *desc=NULL,
-				const char *ret_type=NULL, const char*ret_desc=NULL) {
-		ade_table_entry ate;
-
-		ate.Name = name;
-		ate.Instanced = true;
-		ate.Type = 'v';
-		ate.Value.Function = func;
-		ate.Arguments = args;
-		ate.Description = desc;
-		ate.ReturnType = ret_type;
-		ate.ReturnDescription = ret_desc;
-
-		Ade_table_entries[parent.GetIdx()].AddSubentry(ate);
-		LibIdx = Ade_table_entries.size()-1;
-	}
-};
-
-class ade_indexer : public ade_lib_handle {
-public:
-	ade_indexer(lua_CFunction func, ade_lib_handle &parent, const char *args=NULL, const char *desc=NULL,
-				const char *ret_type=NULL, const char*ret_desc=NULL) {
-		//Add function for meta
-		ade_table_entry ate;
-
-		ate.Name = "__indexer";
-		ate.Instanced = true;
-		ate.Type = 'u';
-		ate.Value.Function = func;
-		ate.Arguments = args;
-		ate.Description = desc;
-		ate.ReturnType = ret_type;
-		ate.ReturnDescription = ret_desc;
-
-		Ade_table_entries[parent.GetIdx()].AddSubentry(ate);
-		LibIdx = Ade_table_entries.size()-1;
-	}
-};
-
-//Struct for converting one string for another. whee!
-struct string_conv {
-	char *src;
-	char *dest;
-};
-
-//*************************Lua operators*************************
-//These are the various types of operators you can
-//set in Lua. Use these as function name to activate.
-//
-//Format string should be "*o" or "o*", where "*" is the type of
-//variable you want to deal with.
-//The order varies with order of variables
-string_conv ade_Operators[] = {
-	{"__add",		"+"},			//var +  obj
-	{"__sub",		"-"},			//var -  obj
-	{"__mul",		"*"},			//var *  obj
-	{"__div",		"/"},			//var /  obj
-	{"__mod",		"%"},			//var %  obj
-	{"__pow",		"^"},			//var ^  obj
-	{"__unm",		"~"},			//var ~  obj
-	{"__concat",	".."},			//var .. obj
-	{"__len",		"#"},			//#var
-	{"__eq",		"=="},			//var == obj
-	{"__lt",		"<"},			//var <  obj
-	{"__le",		"<="},			//var <= obj
-	{"__newindex",	"="},			//var =  obj
-	{"__call",		""},			//*shrug*
-	{"__gc",		"__gc"},		//Lua's equivelant of a destructor
-	//WMC - Used with tostring() scripting operator.
-	{"__tostring",	"(string)"},	//tostring(var)
-	//WMC - This is NOT a Lua type, but for the LUA_INDEXER define
-	{"__indexer",	"[]"},			//obj[var]
-};
-
-int ade_Num_operators = sizeof(ade_Operators)/sizeof(string_conv);
-
-int ade_get_operator(const char *tablename)
-{
-	for(int i = 0; i < ade_Num_operators; i++)
-	{
-		if(!strcmp(tablename, ade_Operators[i].src))
-			return i;
-	}
-
-	return -1;
-}
+using namespace scripting;
 
 const size_t INVALID_ID = (size_t) -1; // Use -1 to get highest possible unsigned number
 
@@ -239,11 +83,6 @@ const size_t INVALID_ID = (size_t) -1; // Use -1 to get highest possible unsigne
 	static int parent##___indexer_f(lua_State *L);		\
 	ade_indexer parent##___indexer(parent##___indexer_f, parent, args, desc, ret_type, ret_desc);	\
 	static int parent##___indexer_f(lua_State *L)
-
-//Checks to determine whether ADE_VIRTVAR or LUA_INDEXER should set the variable
-#define ADE_FUNCNAME_UPVALUE_INDEX	1
-#define ADE_SETTING_UPVALUE_INDEX	2
-#define ADE_SETTING_VAR	lua_toboolean(L,lua_upvalueindex(ADE_SETTING_UPVALUE_INDEX))
 
 //*************************Lua return values*************************
 #define ADE_RETURN_NIL				0
@@ -11968,7 +11807,7 @@ ADE_VIRTVAR(CountermeasureCount, l_Control_Info, "number", "Number of countermea
 	return ade_set_args(L, "i", Player->lua_ci.fire_countermeasure_count);
 }
 
-ADE_FUNC(clearLuaButtonInfo, l_Control_Info, NULL, "Clears the scripting button control info", NULL, NULL)
+ADE_FUNC(clearLuaButtonInfo, l_Control_Info, NULL, "Clears the lua button control info", NULL, NULL)
 {
 	button_info_clear(&Player->lua_bi);
 
@@ -12008,7 +11847,7 @@ ADE_FUNC(accessButtonInfo, l_Control_Info, "number, number, number, number", "Ac
 	return ade_set_args(L, "iiii", bi_status[0], bi_status[1], bi_status[2], bi_status[3]);
 }
 
-ADE_FUNC(useButtonControl, l_Control_Info, "number, string", "Adds the defined button control to scripting button control data, if number is -1 it tries to use the string", NULL, NULL)
+ADE_FUNC(useButtonControl, l_Control_Info, "number, string", "Adds the defined button control to lua button control data, if number is -1 it tries to use the string", NULL, NULL)
 {
 	int index;
 	char *buf = NULL;
@@ -12072,7 +11911,7 @@ ADE_FUNC(getButtonControlNumber, l_Control_Info, "string", "Gives the number of 
 	return ade_set_error(L, "i", -1);
 }
 
-ADE_VIRTVAR(AllButtonPolling, l_Control_Info, "boolean", "Toggles the all button polling for scripting", "boolean", "If the all button polling is enabled or not")
+ADE_VIRTVAR(AllButtonPolling, l_Control_Info, "boolean", "Toggles the all button polling for lua", "boolean", "If the all button polling is enabled or not")
 {
 	bool p;
 	int idx;
@@ -14855,7 +14694,7 @@ ADE_FUNC(__len, l_HookVar_Globals, NULL, "Number of HookVariables", "number", "N
 		total_len++;
 		lua_pop(L, 1);	//value
 	}
-	size_t num_sub = Ade_table_entries[l_HookVar.GetIdx()].Num_subentries;
+	size_t num_sub = ade_manager::getInstance()->getEntry(l_HookVar.GetIdx()).Num_subentries;
 
 	lua_pop(L, 3);
 
@@ -14945,7 +14784,7 @@ ADE_INDEXER(l_Mission_Asteroids, "number Index", "Gets asteroid", "asteroid", "A
 		return ade_set_error( L, "o", l_Asteroid.Set( object_h() ) );
 	}
 	if( idx > -1 && idx < asteroid_count() ) {
-		idx--; //Convert from Lua to C, as scripting indices start from 1, not 0
+		idx--; //Convert from Lua to C, as lua indices start from 1, not 0
 		return ade_set_args( L, "o", l_Asteroid.Set( object_h( &Objects[Asteroids[idx].objnum] ), Objects[Asteroids[idx].objnum].signature ) );
 	}
 
@@ -16225,29 +16064,6 @@ int ade_set_object_with_breed(lua_State *L, int obj_idx)
 lua_Debug Ade_debug_info;
 char debug_stack[4][32];
 
-void ade_debug_call(lua_State *L, lua_Debug *ar)
-{
-	Assert(L != NULL);
-	Assert(ar != NULL);
-	lua_getstack(L, 1, ar);
-	lua_getinfo(L, "nSlu", ar);
-	memcpy(&Ade_debug_info, ar, sizeof(lua_Debug));
-
-	int n;
-	for (n = 0; n < 4; n++) {
-		debug_stack[n][0] = '\0';
-	}
-
-	for (n = 0; n < 4; n++) {
-		if (lua_getstack(L,n+1, ar) == 0)
-			break;
-		lua_getinfo(L,"n", ar);
-		if (ar->name == NULL)
-			break;
-		strcpy_s(debug_stack[n],ar->name);
-	}
-}
-
 void ade_debug_ret(lua_State *L, lua_Debug *ar)
 {
 	Assert(L != NULL);
@@ -16341,11 +16157,11 @@ int script_state::CreateLuaState()
 	//*****INITIALIZE ADE
 	uint i;
 	mprintf(("LUA: Beginning ADE initialization\n"));
-	for(i = 0; i < Ade_table_entries.size(); i++)
+	for(i = 0; i < ade_manager::getInstance()->getNumEntries(); i++)
 	{
 		//WMC - Do only toplevel table entries, doi
-		if(Ade_table_entries[i].ParentIdx == UINT_MAX)			//WMC - oh hey, we're done with the meaty point in < 10 lines.
-			Ade_table_entries[i].SetTable(L, LUA_GLOBALSINDEX, LUA_GLOBALSINDEX);	//Oh the miracles of OOP.
+		if(ade_manager::getInstance()->getEntry(i).ParentIdx == UINT_MAX)			//WMC - oh hey, we're done with the meaty point in < 10 lines.
+			ade_manager::getInstance()->getEntry(i).SetTable(L, LUA_GLOBALSINDEX, LUA_GLOBALSINDEX);	//Oh the miracles of OOP.
 	}
 
 	//*****INITIALIZE RETURN HACK FUNCTION
@@ -16378,1164 +16194,11 @@ void script_state::EndLuaFrame()
 	memcpy(NextDrawStringPos, NextDrawStringPosInitial, sizeof(NextDrawStringPos));
 }
 
-//*************************Lua functions*************************
-//WMC - Spits out the current Lua stack to "stackdump"
-//This includes variable values, but not names
-void ade_stackdump(lua_State *L, char *stackdump)
-{
-	char buf[512];
-	int stacksize = lua_gettop(L);
-
-	//Lua temps
-	double d;
-	int b;
-	char *s;
-	for(int argnum = 1; argnum <= stacksize; argnum++)
-	{
-		int type = lua_type(L, argnum);
-		sprintf(buf, "\r\n%d: ", argnum);
-		strcat(stackdump, buf);
-		switch(type)
-		{
-			case LUA_TNIL:
-				strcat(stackdump, "NIL");
-				break;
-			case LUA_TNUMBER:
-				d = lua_tonumber(L, argnum);
-				sprintf(buf, "Number [%f]",d);
-				strcat(stackdump, buf);
-				break;
-			case LUA_TBOOLEAN:
-				b = lua_toboolean(L, argnum);
-				sprintf(buf, "Boolean [%d]",b);
-				strcat(stackdump, buf);
-				break;
-			case LUA_TSTRING:
-				s = (char *)lua_tostring(L, argnum);
-				sprintf(buf, "String [%s]",s);
-				strcat(stackdump, buf);
-				break;
-			case LUA_TTABLE:
-				{
-					if(lua_getmetatable(L, argnum))
-					{
-						lua_pushstring(L, "__adeid");
-						lua_rawget(L, -2);
-						if(lua_isnumber(L, -1))
-						{
-							sprintf(buf, "Table [%s]", Ade_table_entries[(uint)lua_tonumber(L, -1)].Name);
-							strcat(stackdump, buf);
-						}
-						else
-							strcat(stackdump, "non-default Table");
-						lua_pop(L, 2);	//metatable and nil/adeid
-					}
-					else
-						strcat(stackdump, "Table w/ no metatable");
-
-					//Maybe get first key?
-					char *firstkey = NULL;
-					lua_pushnil(L);
-					if(lua_next(L, argnum))
-					{
-						firstkey = (char *)lua_tostring(L, -2);
-						if(firstkey != NULL)
-						{
-							strcat(stackdump, ", First key: [");
-							strcat(stackdump, firstkey);
-							strcat(stackdump, "]");
-						}
-						lua_pop(L, 1);	//Key
-					}
-					lua_pop(L, 1);	//Nil
-				}
-				break;
-			case LUA_TFUNCTION:
-				strcat(stackdump, "Function");
-				{
-					char *upname = (char*)lua_getupvalue(L, argnum, ADE_FUNCNAME_UPVALUE_INDEX);
-					if(upname != NULL)
-					{
-						strcat(stackdump, " ");
-						strcat(stackdump, lua_tostring(L, -1));
-						strcat(stackdump, "()");
-						lua_pop(L, 1);
-					}
-				}
-				break;
-			case LUA_TUSERDATA:
-				if(lua_getmetatable(L, argnum))
-				{
-					lua_pushstring(L, "__adeid");
-					lua_rawget(L, -2);
-					if(lua_isnumber(L, -1))
-					{
-						sprintf(buf, "Userdata [%s]", Ade_table_entries[(uint)lua_tonumber(L, -1)].Name);
-					}
-					else
-						sprintf(buf, "non-default Userdata");
-
-					lua_pop(L, 2);	//metatable and nil/adeid
-				}
-				else
-					sprintf(buf, "Userdata w/ no metatable");
-				strcat(stackdump, buf);
-				break;
-			case LUA_TTHREAD:
-				sprintf(buf, "Thread");
-				strcat(stackdump, buf);
-				break;
-			case LUA_TLIGHTUSERDATA:
-				sprintf(buf, "Light userdata");
-				strcat(stackdump, buf);
-				break;
-			default:
-				sprintf(buf, "<UNKNOWN>: %s (%f) (%s)", lua_typename(L, type), lua_tonumber(L, argnum), lua_tostring(L, argnum));
-				strcat(stackdump, buf);
-				break;
-		}
-	}
-}
-
-//WMC - Sometimes this gets out of sync between Lua versions
-static const char *Lua_type_names[] = {
-	"nil",
-	"boolean",
-	"light userdata",
-	"number",
-	"string",
-	"table",
-	"function",
-	"userdata",
-	"thread",
-};
-
-static int Lua_type_names_num = sizeof(Lua_type_names)/sizeof(char*);
-
-//WMC - Gets type of object
-const char *ade_get_type_string(lua_State *L, int argnum)
-{
-	int type = lua_type(L, argnum);
-
-	if(type < 0 || type >= Lua_type_names_num)
-		return "Unknown";
-
-	return Lua_type_names[type];
-}
-
-//WMC - hack to skip X number of arguments on the stack
-//Lets me use ade_get_args for global hook return values
-int Ade_get_args_skip = 0;
-bool Ade_get_args_lfunction = false;
-
-//ade_get_args(state, arguments, variables)
-//----------------------------------------------
-//based on "Programming in Lua"
-//
-//Parses arguments from string to variables given
-//a '|' divides required and optional arguments.
-//Returns 0 if a required argument is invalid,
-//or there are too few arguments actually passed
-//
-//NOTE: This function essentially takes objects
-//from the stack in series, so it can easily be used
-//to get the return values from a chunk of Lua code
-//after it has been executed. See RunByteCode()
-int ade_get_args(lua_State *L, const char *fmt, ...)
-{
-	//Check that we have all the arguments that we need
-	//If we don't, return 0
-	int needed_args = (int)strlen(fmt);
-	int total_args = lua_gettop(L) - Ade_get_args_skip;
-
-	if(strchr(fmt, '|') != NULL) {
-		needed_args = (int)(strchr(fmt, '|') - fmt);
-	}
-
-	char funcname[128] = "\0";
-#ifndef NDEBUG
-	lua_Debug ar;
-	memset(&ar, 0, sizeof(ar));
-	if(lua_getstack(L, 0, &ar))
-	{
-		lua_getinfo(L, "nl", &ar);
-		strcpy_s(funcname, "");
-		if(ar.name != NULL) {
-			strcat_s(funcname, ar.name);
-		}
-		if(ar.currentline > -1) {
-			char buf[33];
-			sprintf(buf, "%d", ar.currentline);
-			strcat_s(funcname, " (Line ");
-			strcat_s(funcname, buf);
-			strcat_s(funcname, ")");
-		}
-	}
-#endif
-	if(!strlen(funcname)) {
-		//WMC - This was causing crashes with user-defined functions.
-		//WMC - Try and get at function name from upvalue
-		if(!Ade_get_args_lfunction && !lua_isnone(L, lua_upvalueindex(ADE_FUNCNAME_UPVALUE_INDEX)))
-		{
-				if(lua_type(L, lua_upvalueindex(ADE_FUNCNAME_UPVALUE_INDEX)) == LUA_TSTRING)
-					strcpy_s(funcname, lua_tostring(L, lua_upvalueindex(ADE_FUNCNAME_UPVALUE_INDEX)));
-		}
-
-		//WMC - Totally unknown function
-		if(!strlen(funcname)) {
-			strcpy_s(funcname, "<UNKNOWN>");
-		}
-	}
-	if(total_args < needed_args) {
-		LuaError(L, "Not enough arguments for '%s' - need %d, had %d. If you are using objects or handles, make sure that you are using \":\" to access member functions, rather than \".\"", funcname, needed_args, total_args);
-		return 0;
-	}
-
-	//Start throught
-	va_list vl;
-	int nargs;
-	int counted_args = 0;
-
-	//Are we parsing optional args yet?
-	bool optional_args = false;
-
-	va_start(vl, fmt);
-	nargs = 1 + Ade_get_args_skip;
-	total_args += Ade_get_args_skip;
-	while(*fmt && nargs <= total_args)
-	{
-		//Skip functions; I assume these are being used to return args
-		while(lua_type(L, nargs) == LUA_TFUNCTION && nargs <= total_args)
-			nargs++;
-
-		if(nargs > total_args)
-			break;
-
-		switch(*fmt++)
-		{
-			case 'b':
-				if(lua_isboolean(L, nargs)) {
-					*va_arg(vl, bool*) = lua_toboolean(L, nargs) > 0 ? true : false;
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; boolean expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 'd':
-				if(lua_isnumber(L, nargs)) {
-					*va_arg(vl, double*) = (double)lua_tonumber(L, nargs);
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; number expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 'f':
-				if(lua_isnumber(L, nargs)) {
-					*va_arg(vl, float*) = (float)lua_tonumber(L, nargs);
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; number expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 'i':
-				if(lua_isnumber(L, nargs)) {
-					*va_arg(vl, int*) = (int)lua_tonumber(L, nargs);
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; number expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 's':
-				if(lua_isstring(L, nargs)) {
-					*va_arg(vl, const char **) = lua_tostring(L, nargs);
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; string expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 'x':
-				if(lua_isnumber(L, nargs)) {
-					*va_arg(vl, fix*) = fl2f((float)lua_tonumber(L, nargs));
-				} else {
-					LuaError(L, "%s: Argument %d is an invalid type '%s'; number expected", funcname, nargs, ade_get_type_string(L, nargs));
-					if(!optional_args) {
-						va_end(vl);
-						return 0;
-					}
-				}
-				break;
-			case 'o':
-				{
-					ade_odata od = va_arg(vl, ade_odata);
-					if(lua_isuserdata(L, nargs))
-					{
-						//WMC - Get metatable
-						lua_getmetatable(L, nargs);
-						int mtb_ldx = lua_gettop(L);
-						Assert(!lua_isnil(L, -1));
-	
-						//Get ID
-						lua_pushstring(L, "__adeid");
-						lua_rawget(L, mtb_ldx);
-	
-						if(lua_tonumber(L, -1) != od.idx)
-						{
-							lua_pushstring(L, "__adederivid");
-							lua_rawget(L, mtb_ldx);
-							if((uint)lua_tonumber(L, -1) != od.idx)
-							{
-								LuaError(L, "%s: Argument %d is the wrong type of userdata; '%s' given, but '%s' expected", funcname, nargs, Ade_table_entries[(uint)lua_tonumber(L, -2)].Name, Ade_table_entries[od.idx].GetName());
-								if(!optional_args) {
-									va_end(vl);
-									return 0;
-								}
-							}
-							lua_pop(L, 1);
-						}
-						lua_pop(L, 2);
-						if(od.size != ODATA_PTR_SIZE)
-						{
-							memcpy(od.buf, lua_touserdata(L, nargs), od.size);
-							if(od.sig != NULL) {
-								//WMC - char must be 1
-								Assert(sizeof(char) == 1);
-								//WMC - Yuck. Copy sig data.
-								//Maybe in the future I'll do a packet userdata thing.
-								(*od.sig) = *(ODATA_SIG_TYPE*)(*(char **)od.buf + od.size);
-							}
-						} else {
-							(*(void**)od.buf) = lua_touserdata(L, nargs);
-						}
-					}
-					else if(lua_isnil(L, nargs) && optional_args)
-					{
-						//WMC - Modder has chosen to ignore this argument
-					}
-					else
-					{
-						LuaError(L, "%s: Argument %d is an invalid type '%s'; type '%s' expected", funcname, nargs, ade_get_type_string(L, nargs), Ade_table_entries[od.idx].GetName());
-						if(!optional_args) {
-							va_end(vl);
-							return 0;
-						}
-					}
-				}
-				break;
-			case '|':
-				nargs--;	//cancel out the nargs++ at the end
-				counted_args--;
-				optional_args = true;
-				break;
-			case '*':
-				//WMC - Ignore one spot
-				break;
-			default:
-				Error(LOCATION, "%s: Bad character passed to ade_get_args; (%c)", funcname, *(fmt-1));
-				break;
-		}
-		nargs++;
-		counted_args++;
-	}
-	va_end(vl);
-	return counted_args;
-}
-
-//ade_set_args(state, arguments, variables)
-//----------------------------------------------
-//based on "Programming in Lua"
-//
-//Takes variables given and pushes them onto the
-//Lua stack. Use it to return variables from a
-//Lua scripting function.
-//
-//NOTE: You can also use this to push arguments
-//on to the stack in series. See script_state::SetHookVar
-int ade_set_args(lua_State *L, const char *fmt, ...)
-{
-	//Start throught
-	va_list vl;
-	int nargs;
-	int setargs;	//args actually set
-
-	va_start(vl, fmt);
-	nargs = 0;
-	setargs = 0;
-	while(*fmt != '\0')
-	{
-		switch(*fmt++)
-		{
-			case '*':
-				lua_pushnil(L);
-				break;
-			case 'b':	//WMC - Bool is actually int for GCC (Why...?)
-				lua_pushboolean(L, va_arg(vl, int) ? 1 : 0);
-				break;
-			case 'd':
-				lua_pushnumber(L, va_arg(vl, double));
-				break;
-			case 'f':
-				lua_pushnumber(L, va_arg(vl, double));
-				break;
-			case 'i':
-				lua_pushnumber(L, va_arg(vl, int));
-				break;
-			case 's':
-				{
-					//WMC - Isn't working with HookVar for some strange reason
-					char *s = va_arg(vl, char*);
-					lua_pushstring(L, s);
-					break;
-				}
-			case 'x':
-				lua_pushnumber(L, f2fl(va_arg(vl, fix)));
-				break;
-			case 'o':
-				{
-					//WMC - char must be 1 byte, foo.
-					Assert(sizeof(char)==1);
-					//WMC - step by step
-					//Copy over objectdata
-					ade_odata od = (ade_odata) va_arg(vl, ade_odata);
-
-					//Create new LUA object and get handle
-					char *newod = (char*)lua_newuserdata(L, od.size + sizeof(ODATA_SIG_TYPE));
-					//Create or get object metatable
-					luaL_getmetatable(L, Ade_table_entries[od.idx].Name);
-					//Set the metatable for the object
-					lua_setmetatable(L, -2);
-
-					//Copy the actual object data to the Lua object
-					memcpy(newod, od.buf, od.size);
-
-					//Also copy in the unique sig
-					if(od.sig != NULL)
-						memcpy(newod + od.size, od.sig, sizeof(ODATA_SIG_TYPE));
-					else
-					{
-						ODATA_SIG_TYPE tempsig = ODATA_SIG_DEFAULT;
-						memcpy(newod + od.size, &tempsig, sizeof(ODATA_SIG_TYPE));
-					}
-					break;
-				}
-			//WMC -  Don't forget to update lua_set_arg
-			default:
-				Error(LOCATION, "Bad character passed to ade_set_args; (%c)", *(fmt-1));
-				setargs--;
-		}
-		nargs++;
-		setargs++;
-	}
-	va_end(vl);
-	return setargs;
-}
-
-int ade_friendly_error(lua_State *L)
-{
-	LuaError(L);
-
-	//WMC - According to documentation, this will always be the error
-	//if error handler is called
-	return LUA_ERRRUN;
-}
-
-//WMC - Used to get tostring from object, or just return string pointer
-//if object is already a string.
-char *ade_concat_helper(lua_State *L, int obj_ldx)
-{
-	char *rtn = NULL;
-
-	lua_pushcfunction(L, ade_friendly_error);
-	int err_ldx = lua_gettop(L);
-
-	if(lua_isstring(L, obj_ldx))
-	{
-		rtn = (char*)lua_tostring(L, obj_ldx);
-	}
-	else if(lua_isuserdata(L, obj_ldx))
-	{
-		if(lua_getmetatable(L, obj_ldx))
-		{
-			int mtb_ldx = lua_gettop(L);
-
-			lua_pushstring(L, "__tostring");
-			lua_rawget(L, mtb_ldx);
-
-			if(lua_iscfunction(L, -1))
-			{
-				lua_pushvalue(L, obj_ldx);
-				if(!lua_pcall(L, 1, 1, err_ldx) && lua_type(L, -1) == LUA_TSTRING)
-				{
-					rtn = (char*)lua_tostring(L, -1);
-				}
-			}
-		}
-	}
-
-	//WMC - Clear out all the extra crap.
-	lua_pop(L, lua_gettop(L) - err_ldx + 1);
-
-	return rtn;
-}
-
-//1: Userspace variables (ie in object table)
-//2: Handle-specific values
-//3: Entries in metatable (ie defined by ADE)
-//4: Virtual variables
-//5: Use the indexer, if possible
-//6: Set userspace variable
-//7: Set handle-specific variables
-//X: Mission failed.
-//
-//On the stack when this is called:
-//Index 1 - Object (Can be anything with Lua 5.1; Number to a library)
-//Index 2 - String (ie the key we're trying to access; Object.string, Object:string, Object['string'], etc)
-//Index 3 - (Optional) Argument we are trying to set Object.String = Argument
-static int ade_index_handler(lua_State *L)
-{
-	Assert(L != NULL);
-
-	const int obj_ldx = 1;
-	const int key_ldx = 2;
-	const int arg_ldx = 3;
-	int last_arg_ldx = lua_gettop(L);
-	const char *type_name = NULL;
-	uint ade_id = UINT_MAX;
-	int mtb_ldx = INT_MAX;
-	ade_table_entry* entry = 0;
-
-	//*****STEP 1: Check for user-defined objects
-	if(lua_istable(L, obj_ldx) && !ADE_SETTING_VAR)
-	{
-		lua_pushvalue(L, key_ldx);
-		lua_rawget(L, obj_ldx);
-		if(!lua_isnil(L, -1))
-			return 1;
-		else
-			lua_pop(L, 1);	//nil value
-	}
-
-	//*****STEP 1.5: Set-up metatable
-	if(lua_getmetatable(L, obj_ldx))
-	{
-		mtb_ldx = lua_gettop(L);
-		lua_pushcfunction(L, ade_friendly_error);
-		int err_ldx = lua_gettop(L);
-		int i;
-		
-		//*****WMC - go for the type name
-		lua_pushstring(L, "__adeid");
-		lua_rawget(L, mtb_ldx);
-		if(lua_isnumber(L, -1))
-		{
-			ade_id = (uint) lua_tonumber(L, -1);
-			if(ade_id < Ade_table_entries.size())
-			{
-				entry = &Ade_table_entries[ade_id];
-				type_name = entry->Name;
-			}
-		}
-		lua_pop(L, 1);
-
-		//*****STEP 2: Check for handle signature-specific values
-		if(lua_isuserdata(L, obj_ldx) && ade_id != UINT_MAX && !ADE_SETTING_VAR)
-		{
-			//WMC - I assume char is one byte
-			Assert(sizeof(char) == 1);
-
-			//Get userdata sig
-			char *ud = (char *)lua_touserdata(L, obj_ldx);
-			ODATA_SIG_TYPE sig = *(ODATA_SIG_TYPE*)(ud + entry->Value.Object.size);
-
-			//Now use it to index the table with that #
-			lua_pushnumber(L, sig);
-			lua_rawget(L, mtb_ldx);
-			if(lua_istable(L, -1))
-			{
-				int hvt_ldx = lua_gettop(L);
-				lua_pushvalue(L, key_ldx);
-				lua_rawget(L, hvt_ldx);
-				if(!lua_isnil(L, -1))
-					return 1;
-				else
-					lua_pop(L, 1);	//nil value
-			}
-			lua_pop(L, 1);	//sig table
-		}
-
-		//*****STEP 3: Check for __ademember objects (ie defaults)
-		lua_pushstring(L, "__ademembers");
-		lua_rawget(L, mtb_ldx);
-		if(lua_istable(L, -1))
-		{
-			int amt_ldx = lua_gettop(L);
-			lua_pushvalue(L, key_ldx);
-			lua_rawget(L, amt_ldx);
-			if(!lua_isnil(L, -1))
-				return 1;
-			else
-				lua_pop(L, 1);	//nil value
-		}
-		lua_pop(L, 1);	//member table
-
-		//*****STEP 4: Check for virtual variables
-		lua_pushstring(L, "__virtvars");
-		lua_rawget(L, mtb_ldx);
-		if(lua_istable(L, -1))
-		{
-			//Index virtvar function
-			int vvt_ldx = lua_gettop(L);
-			lua_pushvalue(L, key_ldx);
-			lua_rawget(L, vvt_ldx);
-			if(lua_isfunction(L, -1))
-			{
-				//Set upvalue
-				lua_pushvalue(L, lua_upvalueindex(ADE_SETTING_UPVALUE_INDEX));
-				if(lua_setupvalue(L, -2, ADE_SETTING_UPVALUE_INDEX) == NULL)
-					LuaError(L, "Unable to set upvalue for virtual variable");
-
-				//Set arguments
-				//WMC - Skip setting the key
-				lua_pushvalue(L, obj_ldx);
-				int numargs = 1;
-				for(i = arg_ldx; i <= last_arg_ldx; i++)
-				{
-					lua_pushvalue(L, i);
-					numargs++;
-				}
-
-				//Execute function
-				lua_pcall(L, numargs, LUA_MULTRET, err_ldx);
-
-				return (lua_gettop(L) - vvt_ldx);
-			}
-			else
-			{
-				lua_pop(L, 1);	//non-function value
-			}
-		}
-		lua_pop(L, 1);	//virtvar table
-
-		//*****STEP 5: Use the indexer
-		//NOTE: Requires metatable from step 1.5
-
-		//Get indexer
-		lua_pushstring(L, "__indexer");
-		lua_rawget(L, mtb_ldx);
-		if(lua_isfunction(L, -1))
-		{
-			//Function already on stack
-			//Set upvalue
-			lua_pushvalue(L, lua_upvalueindex(ADE_SETTING_UPVALUE_INDEX));
-			if(lua_setupvalue(L, -2, ADE_SETTING_UPVALUE_INDEX) == NULL)
-				LuaError(L, "Unable to set upvalue for indexer");
-
-			//Set arguments
-			for(i = 1; i <= last_arg_ldx; i++)
-				lua_pushvalue(L, i);
-
-			//Execute function
-			lua_pcall(L, last_arg_ldx, LUA_MULTRET, err_ldx);
-
-			return (lua_gettop(L) - err_ldx);
-		}
-		lua_pop(L, 2);	//WMC - Don't need __indexer or error handler
-	}
-
-	//*****STEP 6: Set a new variable or die.
-	if(ADE_SETTING_VAR && lua_istable(L, obj_ldx))
-	{
-		lua_pushvalue(L, key_ldx);
-		lua_pushvalue(L, arg_ldx);
-		lua_rawset(L, obj_ldx);
-
-		lua_pushvalue(L, key_ldx);
-		lua_rawget(L, obj_ldx);
-		return 1;
-	}
-	//*****STEP 7: Set sig thingie
-	else if(ADE_SETTING_VAR && ade_id != UINT_MAX && mtb_ldx != INT_MAX && lua_isuserdata(L, obj_ldx))
-	{
-		//WMC - I assume char is one byte
-		Assert(sizeof(char) == 1);
-
-		//Get userdata sig
-		char *ud = (char *)lua_touserdata(L, obj_ldx);
-		ODATA_SIG_TYPE sig = *(ODATA_SIG_TYPE*)(ud + entry->Value.Object.size);
-
-		//Now use it to index the table with that #
-		lua_pushnumber(L, sig);
-		lua_rawget(L, mtb_ldx);
-
-		//Create table, if necessary
-		if(!lua_istable(L, -1))
-		{
-			lua_pop(L, 1);
-			lua_newtable(L);
-			lua_pushnumber(L, sig);
-			lua_pushvalue(L, -2);
-			lua_rawset(L, mtb_ldx);
-		}
-
-		//Index the table
-		if(lua_istable(L, -1))
-		{
-			int hvt_ldx = lua_gettop(L);
-			lua_pushvalue(L, key_ldx);
-			lua_pushvalue(L, arg_ldx);
-			lua_rawset(L, hvt_ldx);
-
-			lua_pushvalue(L, key_ldx);
-			lua_rawget(L, hvt_ldx);
-			return 1;
-		}
-		lua_pop(L, 1);	//WMC - maybe-sig-table
-	}
-	lua_pop(L, 1);	//WMC - metatable
-
-	if(type_name != NULL)
-		LuaError(L, "Could not find index '%s' in type '%s'", lua_tostring(L, key_ldx), type_name);
-	else
-		LuaError(L, "Could not find index '%s'", lua_tostring(L, key_ldx));
-	return 0;
-}
-
-//Think of n_mtb_ldx as the parent metatable
-int ade_table_entry::SetTable(lua_State *L, int p_amt_ldx, int p_mtb_ldx)
-{
-	uint i;
-	int cleanup_items = 0;
-	int mtb_ldx = INT_MAX;
-	int data_ldx = INT_MAX;
-	int desttable_ldx = INT_MAX;
-	int amt_ldx = INT_MAX;
-
-	if(Instanced)
-	{
-		//Set any actual data
-		int nset = 0;
-		switch (Type)
-		{
-			//WMC - This hack by taylor is a necessary evil.
-			//64-bit function pointers do not get passed properly
-			//when using va_args for some reason.
-			case 'u':
-			case 'v':
-				lua_pushstring(L, "<UNNAMED FUNCTION>");
-				lua_pushboolean(L, 0);
-				lua_pushcclosure(L, Value.Function, 2);
-				nset++;
-				break;
-
-			default:
-				char typestr[2] = {Type, '\0'};
-				nset = ade_set_args(L, typestr, Value);
-				break;
-		}
-				
-		if (nset)
-		{
-			data_ldx = lua_gettop(L);
-		}
-		else
-		{
-			LuaError(L, "ade_table_entry::SetTable - Could not set data for '%s' (" PTRDIFF_T_ARG ")", GetName(), ADE_INDEX(this));
-		}
-
-		if(data_ldx != INT_MAX)
-		{
-			//WMC - Cannot delete libs and stuff off here.
-			if(p_amt_ldx != LUA_GLOBALSINDEX)
-			{
-				cleanup_items++;
-			}
-
-			//WMC - Handle virtual variables by getting their table
-			if(Type == 'v')
-			{
-				//Get virtvars table
-				lua_pushstring(L, "__virtvars");
-				lua_rawget(L, p_mtb_ldx);
-				if(lua_istable(L, -1))
-				{
-					cleanup_items++;
-
-					//Virtual variables are stored in virtvar table,
-					//rather than the parent table
-					desttable_ldx = lua_gettop(L);
-				}
-				else
-				{
-					lua_pop(L, 1);
-				}
-			}
-			else
-			{
-				//WMC - Member objects prefixed with __ are assumed to be metatable objects
-				if(strnicmp("__", GetName(), 2) && lua_istable(L, p_amt_ldx))
-					desttable_ldx = p_amt_ldx;
-				else if(lua_istable(L, p_mtb_ldx))
-					desttable_ldx = p_mtb_ldx;
-			}
-
-			if(desttable_ldx != INT_MAX)
-			{
-				//If we are setting a function...
-				if(lua_isfunction(L, data_ldx))
-				{
-					//Set the FIRST upvalue to its name,
-					//so we can always find out what it is for debugging
-					lua_pushstring(L, GetName());
-					if(lua_setupvalue(L, data_ldx, 1) == NULL) {
-						LuaError(L, "ade_table_entry::SetTable - Could not set upvalue for '%s' (" PTRDIFF_T_ARG ")", GetName(), ADE_INDEX(this));
-					}
-				}
-
-				//Register name and shortname
-				if(Name != NULL)
-				{
-					lua_pushstring(L, Name);
-					lua_pushvalue(L, data_ldx);
-					lua_rawset(L, desttable_ldx);
-				}
-				if(ShortName != NULL)
-				{
-					lua_pushstring(L, ShortName);
-					lua_pushvalue(L, data_ldx);
-					lua_rawset(L, desttable_ldx);
-				}
-			}
-			else
-			{
-				LuaError(L, "ade_table_entry::SetTable - Could not instance '%s' (" PTRDIFF_T_ARG ")", GetName(), ADE_INDEX(this));
-			}
-		}
-	}
-
-	//If subentries, create a metatable pointer and set it
-	if(Num_subentries || (DerivatorIdx != UINT_MAX && Ade_table_entries[DerivatorIdx].Num_subentries))
-	{
-		//Create the new metatable
-		if(!luaL_newmetatable(L, Name))
-		{
-			LuaError(L, "ade_table_entry::SetTable - Couldn't create metatable for table entry '%s' - does a Lua object already exist with this name?", Name);
-			return 0;
-		}
-		mtb_ldx = lua_gettop(L);
-		cleanup_items++;
-
-		//Push a copy of the metatable and set it for this object
-		//WMC - Make sure it's instanced, too. This helps keep crashes from happening...
-		if(data_ldx != INT_MAX)
-		{
-			lua_pushvalue(L, mtb_ldx);
-			lua_setmetatable(L, data_ldx);
-		}
-
-		//***Create index handler entry
-		lua_pushstring(L, "__index");
-		lua_pushstring(L, "ade_index_handler(get)");	//upvalue(1) = function name
-		lua_pushboolean(L, 0);							//upvalue(2) = setting true/false
-		lua_pushcclosure(L, ade_index_handler, 2);
-		lua_rawset(L, mtb_ldx);
-
-		//***Create newindex handler entry
-		lua_pushstring(L, "__newindex");
-		lua_pushstring(L, "ade_index_handler(set)");	//upvalue(1) = function name
-		lua_pushboolean(L, 1);							//upvalue(2) = setting true/false
-		lua_pushcclosure(L, ade_index_handler, 2);
-		lua_rawset(L, mtb_ldx);
-
-		//***Create virtvar storage facility
-		lua_pushstring(L, "__virtvars");
-		lua_newtable(L);
-		lua_rawset(L, mtb_ldx);
-
-		//***Create ade members table
-		lua_createtable(L, 0, (int)Num_subentries);
-		if(lua_istable(L, -1)) {
-			//WMC - was lua_gettop(L) - 1 for soem
-			amt_ldx = lua_gettop(L);
-			cleanup_items++;
-
-			//Set it
-			lua_pushstring(L, "__ademembers");
-			lua_pushvalue(L, amt_ldx);	//dup
-			lua_rawset(L, mtb_ldx);
-		}
-
-		//***Create ID entries
-		lua_pushstring(L, "__adeid");
-		lua_pushnumber(L, static_cast<lua_Number>(ADE_INDEX(this)));
-		lua_rawset(L, mtb_ldx);
-
-		if(DerivatorIdx != UINT_MAX)
-		{
-			lua_pushstring(L, "__adederivid");
-			lua_pushinteger(L, DerivatorIdx);
-			lua_rawset(L, mtb_ldx);
-		}
-	}
-
-	if(amt_ldx != INT_MAX)
-	{
-		//Fill out ze metatable
-		if(DerivatorIdx != UINT_MAX)
-		{
-			for(i = 0; i < Ade_table_entries[DerivatorIdx].Num_subentries; i++)
-			{
-				Ade_table_entries[Ade_table_entries[DerivatorIdx].Subentries[i]].SetTable(L, amt_ldx, mtb_ldx);
-			}
-		}
-		for(i = 0; i < Num_subentries; i++)
-		{
-			Ade_table_entries[Subentries[i]].SetTable(L, amt_ldx, mtb_ldx);
-		}
-	}
-
-	//Pop the metatable and data (cleanup)
-	lua_pop(L, cleanup_items);
-
-	return 1;
-}
-
-void ade_output_type_link(FILE *fp, const char *typestr)
-{
-	for(int i = 0; i < Lua_type_names_num; i++)
-	{
-		if(!stricmp(Lua_type_names[i], typestr))
-		{
-			fputs(typestr, fp);
-			return;
-		}
-	}
-
-	fprintf(fp, "<a href=\"#%s\">%s</a>", typestr, typestr);
-}
-
-void ade_table_entry::OutputMeta(FILE *fp)
-{
-	if(Name == NULL && ShortName == NULL) {
-		Warning(LOCATION, "Data entry with no name or shortname");
-		return;
-	}
-
-	uint i;
-	bool skip_this = false;
-
-	//WMC - Hack
-	if(ParentIdx != UINT_MAX)
-	{
-		for(i = 0; i < Num_subentries; i++)
-		{
-			if(!stricmp(Ade_table_entries[Subentries[i]].Name, "__indexer"))
-			{
-				Ade_table_entries[Subentries[i]].OutputMeta(fp);
-				skip_this = true;
-				break;
-			}
-		}
-	}
-
-
-	//***Begin entry
-	if(!skip_this)
-	{
-		fputs("<dd><dl>", fp);
-
-		switch(Type)
-		{
-			case 'o':
-				{
-					//***Name (ShortName)
-					if(ParentIdx == UINT_MAX) {
-						fprintf(fp, "<dt id=\"%s\">", GetName());
-					}
-					if(Name == NULL)
-					{
-						if(DerivatorIdx == UINT_MAX)
-							fprintf(fp, "<h2>%s</h2>\n", ShortName);
-						else
-							fprintf(fp, "<h2>%s:<a href=\"#%s\">%s</a></h2>\n", Ade_table_entries[DerivatorIdx].GetName(), Ade_table_entries[DerivatorIdx].GetName(), Ade_table_entries[DerivatorIdx].GetName());
-					}
-					else
-					{
-						fprintf(fp, "<h2>%s", Name);
-
-						if(ShortName != NULL)
-							fprintf(fp, " (%s)", ShortName);
-						if(DerivatorIdx != UINT_MAX)
-							fprintf(fp, ":<a href=\"#%s\">%s</a>", Ade_table_entries[DerivatorIdx].GetName(), Ade_table_entries[DerivatorIdx].GetName());
-
-						fputs("</h2>\n", fp);
-					}
-					fputs("</dt>\n", fp);
-
-					//***Description
-					if(Description != NULL) {
-						fprintf(fp, "<dd>%s</dd>\n", this->Description);
-					}
-
-					//***Type: ReturnType
-					if(ReturnType != NULL) {
-						fprintf(fp, "<dd><b>Return Type: </b> %s<br>&nbsp;</dd>\n", ReturnType);
-					}
-
-					if(ReturnDescription != NULL) {
-						fprintf(fp, "<dd><b>Return Description: </b> %s<br>&nbsp;</dd>\n", ReturnDescription);
-					}
-				}
-				break;
-			case 'u':
-				{
-					//***Name(ShortName)(Arguments)
-					fputs("<dt>", fp);
-					int ao = -1;
-
-					fputs("<i>", fp);
-					if(ReturnType != NULL)
-						ade_output_type_link(fp, ReturnType);
-					else
-						fputs("nil", fp);
-					fputs(" </i>", fp);
-					if(Name == NULL) {
-						fprintf(fp, "<b>%s", ShortName);
-					}
-					else
-					{
-						ao = ade_get_operator(Name);
-
-						//WMC - Do we have an operator?
-						if(ao < 0)
-						{
-							fprintf(fp, "<b>%s", Name);
-							if(ShortName != NULL)
-							{
-								int ao2 = ade_get_operator(ShortName);
-								if(ao2 < 0)
-									fprintf(fp, "(%s)", ShortName);
-								else
-									fprintf(fp, "(%s)", ade_Operators[ao2].dest);
-							}
-						}
-						else
-						{
-							//WMC - Hack
-							if(ParentIdx != UINT_MAX && Ade_table_entries[ParentIdx].ParentIdx != UINT_MAX && Ade_table_entries[ParentIdx].Name != NULL && !stricmp(Name, "__indexer"))
-							{
-								fprintf(fp, "<b>%s%s", Ade_table_entries[ParentIdx].Name, ade_Operators[ao].dest);
-							}
-							else
-								fprintf(fp, "<b>%s", ade_Operators[ao].dest);
-						}
-					}
-					if(ao < 0)
-					{
-						if(Arguments != NULL) {
-							fprintf(fp, "(</b><i>%s</i><b>)</b>\n", Arguments);
-						} else {
-							fprintf(fp, "()</b>\n");
-						}
-					}
-					else
-					{
-						fputs("</b>", fp);
-						if(Arguments != NULL) {
-							fprintf(fp, " <i>%s</i>\n", Arguments);
-						}
-					}
-					fputs("</dt>\n", fp);
-
-					//***Description
-					if(Description != NULL) {
-						fprintf(fp, "<dd>%s</dd>\n", Description);
-					}
-
-					//***Result: ReturnDescription
-					if(ReturnDescription != NULL) {
-						fprintf(fp, "<dd><b>Returns:</b> %s<br>&nbsp;</dd>\n", ReturnDescription);
-					} else {
-						fputs("<dd><b>Returns:</b> Nothing<br>&nbsp;</dd>\n", fp);
-					}
-				}
-				break;
-			default:
-				Warning(LOCATION, "Unknown type '%c' passed to ade_table_entry::OutputMeta", Type);
-			case 'b':
-			case 'd':
-			case 'f':
-			case 'i':
-			case 's':
-			case 'x':
-			case 'v':
-				{
-					//***Type Name(ShortName)
-					fputs("<dt>\n", fp);
-					if(ReturnType != NULL)
-					{
-						fputs("<i>", fp);
-						ade_output_type_link(fp, ReturnType);
-						fputs(" </i>", fp);
-					}
-
-					if(Name == NULL) {
-						fprintf(fp, "<b>%s</b>\n", ShortName);
-					}
-					else
-					{
-						fprintf(fp, "<b>%s", Name);
-						if(ShortName != NULL)
-							fputs(ShortName, fp);
-						fputs("</b>\n", fp);
-					}
-					if(Arguments != NULL)
-						fprintf(fp, " = <i>%s</i>", Arguments);
-					fputs("</dt>\n", fp);
-
-					//***Description
-					if(Description != NULL)
-						fprintf(fp, "<dd>%s</dd>\n", Description);
-
-					//***Also settable with: Arguments
-					if(ReturnDescription != NULL)
-						fprintf(fp, "<dd><b>Value:</b> %s</b></dd>\n", ReturnDescription);
-				}
-				break;
-		}
-	}
-
-	fputs("<dd><dl>\n", fp);
-	for(i = 0; i < Num_subentries; i++)
-	{
-		if(ParentIdx == UINT_MAX
-			|| stricmp(Ade_table_entries[Subentries[i]].Name, "__indexer"))
-			Ade_table_entries[Subentries[i]].OutputMeta(fp);
-	}
-	fputs("</dl></dd>\n", fp);
-
-	if(!skip_this)
-		fputs("<br></dl></dd>\n", fp);
-}
-
 void ade_output_toc(FILE *fp, ade_table_entry *ate)
 {
 	Assert(fp != NULL);
 	Assert(ate != NULL);
-		
+
 	//WMC - sanity checking
 	if(ate->Name == NULL && ate->ShortName == NULL) {
 		Warning(LOCATION, "Found ade_table_entry with no name or shortname");
@@ -17570,9 +16233,9 @@ void script_state::OutputLuaMeta(FILE *fp)
 
 	//***TOC: Libraries
 	fputs("<dt><b>Libraries</b></dt>\n", fp);
-	for(i = 0; i < Ade_table_entries.size(); i++)
+	for(i = 0; i < ade_manager::getInstance()->getNumEntries(); i++)
 	{
-		ate = &Ade_table_entries[i];
+		ate = &ade_manager::getInstance()->getEntry(i);
 		if(ate->ParentIdx == UINT_MAX && ate->Type == 'o' && ate->Instanced) {
 			ade_output_toc(fp, ate);
 		}
@@ -17580,9 +16243,9 @@ void script_state::OutputLuaMeta(FILE *fp)
 
 	//***TOC: Objects
 	fputs("<dt><b>Types</b></dt>\n", fp);
-	for(i = 0; i < Ade_table_entries.size(); i++)
+	for(i = 0; i < ade_manager::getInstance()->getNumEntries(); i++)
 	{
-		ate = &Ade_table_entries[i];
+		ate = &ade_manager::getInstance()->getEntry(i);
 		if(ate->ParentIdx == UINT_MAX && ate->Type == 'o' && !ate->Instanced) {
 			ade_output_toc(fp, ate);
 		}
@@ -17596,9 +16259,9 @@ void script_state::OutputLuaMeta(FILE *fp)
 
 	//***Everything
 	fputs("<dl>\n", fp);
-	for(i = 0; i < Ade_table_entries.size(); i++)
+	for(i = 0; i < ade_manager::getInstance()->getNumEntries(); i++)
 	{
-		ate = &Ade_table_entries[i];
+		ate = &ade_manager::getInstance()->getEntry(i);
 		if(ate->ParentIdx == UINT_MAX)
 			ate->OutputMeta(fp);
 	}
