+++ code/scripting/scripting.cpp
@@ -11,10 +11,13 @@
 #include "io/key.h"
 #include "mission/missioncampaign.h"
 #include "parse/parselo.h"
-#include "scripting.h"
+#include "scripting/scripting.h"
+#include "scripting/ade_args.h"
 #include "ship/ship.h"
 #include "weapon/beam.h"
 #include "weapon/weapon.h"
+
+using namespace scripting;
 
 //tehe. Declare the main event
 script_state Script_system("FS2_Open Scripting");
@@ -972,7 +975,7 @@ void script_state::Clear()
 	LuaLibs = NULL;
 }
 
-script_state::script_state(char *name)
+script_state::script_state(const char *name)
 {
 	strncpy(StateName, name, sizeof(StateName)-1);
 
@@ -980,13 +983,6 @@ script_state::script_state(char *name)
 
 	LuaState = NULL;
 	LuaLibs = NULL;
-}
-
-script_state& script_state::operator=(script_state &in)
-{
-	Error(LOCATION, "SESSION COPY ATTEMPTED");
-
-	return *this;
 }
 
 script_state::~script_state()
@@ -1189,7 +1185,7 @@ void script_state::ParseChunkSub(int *out_lang, int *out_index, char* debug_str)
 		CFILE *cfp = cfopen(filename, "rb", CFILE_NORMAL, CF_TYPE_SCRIPTS );
 		if(cfp == NULL)
 		{
-			Warning(LOCATION, "Could not load scripting script file '%s'", filename);
+			Warning(LOCATION, "Could not load lua script file '%s'", filename);
 		}
 		else
 		{
@@ -1262,7 +1258,7 @@ void script_state::ParseChunkSub(int *out_lang, int *out_index, char* debug_str)
 	{
 		char buf[PARSE_BUF_SIZE];
 
-		//Assume scripting
+		//Assume lua
 		*out_lang = SC_LUA;
 
 		strcpy_s(buf, "return ");
