+++ code/source_groups.cmake
@@ -907,6 +907,10 @@ set (file_root_render
 )
 
 set(file_root_scripting
+	scripting/ade.cpp
+	scripting/ade.h
+	scripting/ade_args.cpp
+	scripting/ade_args.h
 	scripting/lua.cpp
 	scripting/lua.h
 	scripting/scripting.cpp
